requirements.txt
python-telegram-bot~=22.00
PyYAML~=6.0.2
pydantic
